Handle errors within AppRunner
